requirements.txt
Flask==3.1.3
gunicorn==25.3.0
openai==2.30.0
python-dotenv==1.2.2
joblib==1.5.3
scikit-learn==1.7.2
numpy==2.2.6